README.md
Code 1 is the code for model performance evaluation. Here, only the code for the model in dataset 1 is given. To run datasets 2 and 3, you only need to replace the address of dataset 1 with that of dataset 2 and 3.
Code 2 is the code for selecting model parameters, using the grid search algorithm. Here, only the running code of dataset 1 is given, and datasets 2 and 3 only need to rewrite the address in the corresponding position.
For the dataset folder, there are three datasets, and each dataset file contains three files. Sm_s represents the chemical structure similarity data information of drugs, sm_v represents the virus-drug association information, and v_s represents the sequence similarity data information of drugs.
The Accession-number-all 3 datasets folder contains three subfolders, each containing the Accession-number of the virus and the drug respectively.